Fixes the build for mingw-i686 in which the size_t != unsigned long

+++ algorithm.hpp
@@ -84,12 +84,12 @@ void remove_element(T& x, U const y){
 
 template <typename Container>
 auto random_element(Container const& c) -> decltype(c[0]){
-	return c[random_value(0lu, c.size())];
+	return c[random_value(size_t(0), c.size())];
 }
 
 template <typename Container>
 auto random_element(Container& c) -> decltype(c[0]){
-	return c[random_value(0lu, c.size())];
+	return c[random_value(size_t(0), c.size())];
 }
 
 }
